refactor: console용 view로직이 제거된 ChessGame을 domain 패키지로 이동
