+++ src/main/java/chess/domain/chessgame/Camp.java
@@ -1,4 +1,4 @@
-package chess.domain;
+package chess.domain.chessgame;
 
 import java.util.Arrays;
 import java.util.Locale;
