+++ src/main/java/chess/domain/command/Start.java
@@ -1,6 +1,6 @@
 package chess.domain.command;
 
-import chess.console.ChessGame;
+import chess.domain.chessgame.ChessGame;
 
 public final class Start implements CommandGenerator {
     @Override
