+++ src/main/java/chess/domain/piece/Knight.java
@@ -10,10 +10,10 @@ import static chess.domain.board.UnitDirectVector.TOP_TOP_LEFT;
 import static chess.domain.board.UnitDirectVector.TOP_TOP_RIGHT;
 import static chess.domain.piece.PieceProperty.KNIGHT;
 
-import chess.domain.Camp;
 import chess.domain.board.Position;
 import chess.domain.board.Positions;
 import chess.domain.board.UnitDirectVector;
+import chess.domain.chessgame.Camp;
 import java.util.List;
 import java.util.function.Consumer;
