+++ src/main/java/chess/domain/piece/NullPiece.java
@@ -2,10 +2,10 @@ package chess.domain.piece;
 
 import static chess.domain.piece.PieceProperty.NULL_PIECE;
 
-import chess.domain.Camp;
 import chess.domain.board.Position;
 import chess.domain.board.Positions;
 import chess.domain.board.UnitDirectVector;
+import chess.domain.chessgame.Camp;
 import java.util.Collections;
 import java.util.List;
 import java.util.function.Consumer;
