+++ src/main/java/chess/domain/piece/Queen.java
@@ -10,10 +10,10 @@ import static chess.domain.board.UnitDirectVector.TOP_LEFT;
 import static chess.domain.board.UnitDirectVector.TOP_RIGHT;
 import static chess.domain.piece.PieceProperty.QUEEN;
 
-import chess.domain.Camp;
 import chess.domain.board.Position;
 import chess.domain.board.Positions;
 import chess.domain.board.UnitDirectVector;
+import chess.domain.chessgame.Camp;
 import java.util.List;
 import java.util.function.Consumer;
